backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.1
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.6.0
